Refactor forum checkAccess
