+++ src/back/External/Forum/AccessCheck.php
@@ -4,12 +4,195 @@ declare(strict_types=1);
 
 namespace KeepersTeam\Webtlo\External\Forum;
 
+use GuzzleHttp\Exception\GuzzleException;
+use KeepersTeam\Webtlo\Helper;
+
 /**
- * Ошибки при проверке доступа к форуму.
+ * Авторизация на форуме и проверка наличия доступа.
  */
-enum AccessCheck: string
+trait AccessCheck
 {
-    case NOT_AUTHORIZED   = 'Нет доступа на форум. Пройдите авторизацию в настройках.';
-    case USER_CANDIDATE   = 'Нет доступа в рабочий подфорум хранителей. Если вы Кандидат, то ожидайте включения в основную группу.';
-    case VERSION_OUTDATED = 'Отправка отчётов для текущей версии web-TLO заблокирована. Установите актуальную версию web-TLO для корректной отправки отчётов.';
+    /**
+     * Проверка доступности форума.
+     *
+     * @return bool статус доступности форума
+     */
+    public function checkAccess(): bool
+    {
+        // Сперва проверим, что доступ к форуму вообще есть.
+        $availability = $this->checkAvailability();
+        if (!$availability) {
+            $this->logger->error('Не удалось подключиться к форуму. Проверьте настройки доступа в настройках.');
+
+            return false;
+        }
+
+        // Теперь проверим статус авторизации.
+        if ($this->cookie->count() > 0) {
+            $this->logger->debug('Detected non-empty cookie jar, trying stored session first');
+            $isLoggedIn = $this->isLoggedIn();
+
+            if (!$isLoggedIn) {
+                $this->logger->debug('Looks like cookies are rotten, logging in');
+                $loggedIn = $this->autoLogin();
+            } else {
+                $this->logger->debug('Cookies are still fresh');
+                $loggedIn = true;
+            }
+        } else {
+            $this->logger->debug('No cookies found, proceeding with fresh login');
+            $loggedIn = $this->autoLogin();
+        }
+
+        if (!$loggedIn) {
+            $this->logger->error('Не удалось авторизоваться на форуме. Пройдите авторизацию в настройках.');
+
+            return false;
+        }
+
+        return true;
+    }
+
+    /**
+     * Авторизация по кнопке из настроек.
+     *
+     * @param ?array<string, string> $captcha Заполненные коды CAPTCHA
+     *
+     * @return ?Captcha Объект CAPTCHA или null в случае успеха авторизации
+     */
+    public function manualLogin(?array $captcha): ?Captcha
+    {
+        $page = $this->fetchLoginPage(captcha: $captcha);
+
+        if (!empty($page)) {
+            // После авторизации, пробуем получить токен.
+            $token = self::parseFormToken(page: $page);
+
+            // Если токена нет, значит что-то не так.
+            if ($token === null) {
+                return self::parseCaptchaCodes(authPage: $page, logger: $this->logger);
+            }
+
+            // Проверяем новые куки.
+            $this->parseCookie();
+        }
+
+        return null;
+    }
+
+    /**
+     * Получение нового cookie авторизации.
+     *
+     * @return string Новый cookie авторизации или пустая строка
+     */
+    public function getUpdatedCookie(): string
+    {
+        return $this->updatedCookie ?? '';
+    }
+
+    /**
+     * Проверим доступность форума.
+     */
+    private function checkAvailability(): bool
+    {
+        try {
+            $response = $this->client->get(
+                uri    : self::accessURL,
+                options: ['max_retry_attempts' => 1]
+            );
+
+            return $response->getStatusCode() === 200;
+        } catch (GuzzleException) {
+        }
+
+        return false;
+    }
+
+    /**
+     * Проверка, авторизован ли пользователь на форуме.
+     *
+     * @return bool Статус авторизации пользователя
+     */
+    private function isLoggedIn(): bool
+    {
+        // Проверяем доступность папки входящих сообщений.
+        $page = $this->get(
+            url   : self::inboxURL,
+            params: [
+                'allow_redirects' => false, // 302 only in case of user logged out.
+                'query'           => ['folder' => 'inbox'],
+            ]
+        );
+
+        // Если пустой ответ, значит нужно авторизоваться.
+        if (empty($page)) {
+            return false;
+        }
+
+        // Если доступ есть, получаем токен.
+        self::parseFormToken(page: $page);
+
+        return true;
+    }
+
+    /**
+     * Автоматическая авторизация на форуме, используя существующие ключи.
+     *
+     * @return bool статус авторизации
+     */
+    private function autoLogin(): bool
+    {
+        $page = $this->fetchLoginPage();
+        if (!empty($page)) {
+            // После авторизации, пробуем получить токен.
+            $token = self::parseFormToken(page: $page);
+            if ($token === null) {
+                return false;
+            }
+
+            // Проверяем новые куки.
+            $this->parseCookie();
+
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * Получение страницы авторизации.
+     *
+     * @param ?array<string, string> $captcha Заполненные коды CAPTCHA
+     *
+     * @return ?string HTML-страница авторизации или null в случае ошибки
+     */
+    private function fetchLoginPage(?array $captcha = null): ?string
+    {
+        // Очищаем куки авторизации, если есть
+        $this->cookie->clear(name: self::$authCookieName);
+
+        $form = [
+            'login_username' => Helper::encodeCyrillicString($this->cred->auth->username),
+            'login_password' => Helper::encodeCyrillicString($this->cred->auth->password),
+            'login'          => Helper::encodeCyrillicString(self::loginAction),
+        ];
+        if ($captcha !== null) {
+            $form += $captcha;
+        }
+
+        return $this->post(url: self::loginURL, params: ['form_params' => $form]);
+    }
+
+    /**
+     * Получение cookie после авторизации.
+     */
+    private function parseCookie(): void
+    {
+        $cookies = $this->cookie;
+        if ($cookies->count() > 0) {
+            $cookie = $cookies->getCookieByName(name: self::$authCookieName);
+
+            $this->updatedCookie = (string) $cookie;
+        }
+    }
 }
